tombola: Offset the timer instead of steps_executed
When the physics simulation is lagging behind, instead of updating
steps_executed to jump forward in time the elapsed time value is offset
instead. Otherwise the jump in the steps_executed value will make the
tombola suddenly turn really fast and the balls will get a lot of energy
and can escape out of the hexagon. This also avoids the need to have the
WaitingToStart spin state because steps_executed now always increments
only once at each step.